+++ src/lib/middleware/rsc/worker-api.ts
@@ -10,7 +10,14 @@ import type {
 } from "../../../server.js";
 
 const worker = new Worker(new URL("worker-impl.js", import.meta.url), {
-  execArgv: ["--conditions", "react-server"],
+  execArgv: [
+    "--experimental-loader",
+    "waku/node-loader",
+    "--experimental-loader",
+    "react-server-dom-webpack/node-loader",
+    "--conditions",
+    "react-server",
+  ],
 });
 
 export type BuildOutput = {