+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Changed
 - fix: partially for css modules #98
+- feat: use node-loader again #103
 
 ## [0.13.0] - 2023-07-06
 ### Changed